refactor: 강의 기간 객체 분리
